projects/litrelab/litreelab-backend/pyproject.toml
[project]
name = "litreelab-backend"
version = "0.1.0"
description = "Litreelab FastAPI Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "PyJWT[crypto]>=2.8.0",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "openai>=2.28.0",
    "azure-identity>=1.25.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
